+++ siumai/tests/public_surface_imports_test.rs
@@ -482,6 +482,9 @@ fn public_surface_unified_imports_compile() {
     let _ = size_of::<*const dyn ImageModelV4>();
     let _ = size_of::<*const dyn EmbeddingModel>();
     let _ = size_of::<*const dyn LanguageModel>();
+    let _ = size_of::<*const dyn LanguageModelV4>();
+    let _ = size_of::<LanguageModelV4Stream>();
+    let _ = size_of::<LanguageModelV4DoStreamResult>();
     let _ = size_of::<*const dyn LanguageModelMiddleware>();
     let _ = size_of::<*const dyn RerankingModel>();
     let _ = size_of::<*const dyn SpeechModel>();